refactor:#21 - extract heatmap constants and types for reusability
- constants: DAY_LABELS, TIME_LABELS
- interfaces: TimeSlotKey, TimeSlot, DayData, WeeklyHeatmapResponse
